chore: add pagination for export Component&Theme and centralize getUiBuilderService
